Revert "Camera: Switch usage flags to 64-bit unsigned"
This reverts commit 050f5dcf7d1903926b165ebdd2dd30f7a2e83c02.

+++ services/camera/libcameraservice/device1/CameraHardwareInterface.cpp
@@ -382,7 +382,7 @@ CameraHardwareInterface::setUsage(hardware::graphics::common::V1_0::BufferUsage 
         ALOGE("%s: preview window is null", __FUNCTION__);
         return s;
     }
-    mPreviewUsage = static_cast<uint64_t> (usage);
+    mPreviewUsage = (int) usage;
     int rc = native_window_set_usage(a, mPreviewUsage);
     if (rc == OK) {
         cleanupCirculatingBuffers();
